+++ packages/orq-rc/RELEASES.md
@@ -3399,3 +3399,13 @@ Based on:
 - [python v3.14.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.35] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.35 - packages/orq-rc
+
+## 2025-10-20 05:24:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.0 (2.727.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.36] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.36 - packages/orq-rc